% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R, R/praxi.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,crops.class,data.frame-method}
\alias{plot,praxi.class-method}
\title{Visualisation of data, costs, penalty values and anomaly locations.}
\usage{
\S4method{plot}{crops.class}(x)

\S4method{plot}{praxi.class}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{praxi}}.}

\item{y}{A dataframe containing the locations and values of the data points. The data plot is plotted below, and is aligned with, the changepoint locations.}
}
\value{
A ggplot object. Note - if no changepoints are detected in the penalty interval [beta_min,beta_max], then the value returned is NULL.

A ggplot object.
}
\description{
Plot methods for an S4 object returned by \code{\link{crops}}. The plot can also be combined with the original data if required.

A plot method for an S4 object returned by \code{\link{praxi}}.
}
\examples{
# see the crops example

}
\seealso{
\code{\link{crops}}.
}
