% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praxi.R
\name{ouralg}
\alias{ouralg}
\alias{praxi}
\title{Detects point and collective anomalies in correlated time series data.}
\usage{
praxi(y, p, b = NULL)
}
\arguments{
\item{y}{A numeric vector of observations..}

\item{b}{A positive numeric constant for the penalty.}

\item{x}{A positive integer containing the order of the AR model.}
}
\value{
An instance of an S4 class of type \code{\link{praxi}}.
}
\description{
Runs ouralg on a vector of observations.
}
\examples{
library(praxi)

#simulate an AR(1) time series with two changes in mean

gam <- c(0.7); n <- 10000
y <- arima.sim(list(ar=gam),n=n)
y[300] <- y[300]+10
y[5000:5100] <- y[5000:5100]+2

#using custom penalty
res <- praxi(y=y,p=length(gam),b=5*log(n))
#using default penalty
res <- praxi(y=y,p=length(gam))
anoms <- anomalies(res)
pt <- plot(res)
print(res)


}
