% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praxi.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,praxi.class-method}
\title{Produces a plot showing anomalies detected by ouralg.}
\usage{
\S4method{summary}{praxi.class}(object)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{praxi}}.}
}
\value{
A numeric matrix containing the respective positions and change in mean of each detected anomaly.
}
\description{
A summary method for an S4 object returned by \code{\link{praxi}}.
}
